feat: add zero-error pyright gate with strict config, pinned stubs, and error-level fixes
Commit a strict [tool.pyright] configuration (correctness families at
error, noise families as a visible warning backlog), pin pandas-stubs,
scipy-stubs, pyright 1.1.408, and unconditional tomli in the dev extra,
fix all error-level diagnostics across the limen/ tree with
behavior-preserving edits, and gate every PR through a new
pr_checks_pyright workflow. Baseline at this commit: 0 errors,
2,209 warnings.

Closes #650. Parent: #649.

+++ limen/metrics/safe_ovr_auc.py
@@ -35,4 +35,4 @@ def safe_ovr_auc(y_true: np.ndarray, probs: np.ndarray) -> float:
         column_index = _probability_column_index(c, present, probs.shape[1])
         if pos.any() and neg.any() and column_index is not None:
             aucs.append(roc_auc_score(pos, probs[:, column_index]))
-    return float('nan') if not aucs else np.mean(aucs)
+    return float('nan') if not aucs else float(np.mean(aucs))
